Update Dockerfile
- Include Drivers books
- Skip if a book.json can't be retrieved (e.g. 404 because a book doesn't exist in an older version)

curl downloaded a file with content `404: Not Found` before if it didn't exist, which would let `gitbook install` fail because of invalid JSON - which should be harmless because the exit code isn't returned to docker, but we can skip to save some time and print meaningful status messages.

Coupling GSEARCH_ID_* with $book doesn't seem like a good idea as we might replace Google CSE at some point. The versions to download the Gitbook plugins for need to be adjusted manually anyway. Aother option would be to add additional entries to the VERSIONS file, but it would place even more version-independent information into version branches.